Renaming Directory from abstract_descriptor to descriptor
