+++ examples/slottable_data_descriptor_example.py
@@ -1,4 +1,4 @@
-from abstract_descriptor import (
+from descriptor import (
     BaseSlottableDataDescriptor,
     SlottableDefaultDescriptor,
 )
